fix query model

+++ server/src/model/Staff.java
@@ -139,7 +139,7 @@ public class Staff {
         return salary;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(float salary) {
         this.salary = salary;
     }
